requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Reuploaded==1.1.0
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.0
Werkzeug==2.0.1
WTForms==2.3.3
